folder file refactor to allow reuse of form data
